add ai string flags and minor html/css changes
What I call AI flags are pieces of text that can be found on some ai pins. They can be written by authors or found in the comment section to indicate that the content was made by ai. This way, the ai detection does not only rely on Pinterest's detection